requirements.txt
# Core ML Dependencies
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.24.3
matplotlib==3.8.2
seaborn==0.13.0

# MLOps and Experiment Tracking
mlflow==2.8.1

# Google Cloud Dependencies
google-cloud-storage==2.10.0
google-cloud-run==0.10.1
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# API and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Testing Dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Deployment and Containerization
docker==6.1.3

# Monitoring and Observability
prometheus-client==0.19.0
google-cloud-monitoring==2.16.0

# Data Validation and Quality
scipy==1.11.4
great-expectations==0.18.5

# Configuration and Environment
pyyaml==6.0.1
python-dotenv==1.0.0

# HTTP Requests and APIs
requests==2.31.0
httpx==0.25.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development Tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Jupyter Notebook Support (for Colab)
jupyter==1.0.0
ipykernel==6.26.0
ipywidgets==8.1.1

# Email and Notifications
secure-smtplib==0.1.1

# Data Processing
openpyxl==3.1.2
xlsxwriter==3.1.9

# Environment Variable Management
python-decouple==3.8